chore: address issue #618
